Moved music page specific logic to it's own app logic
