+++ rsbuild.config.ts
@@ -30,6 +30,10 @@ export default function (env: any = {}, argv: Record<string, any> = {}) {
           import: './src/bandcamp/content/app.all.ts',
           html: false,
         },
+        'bandcamp.content.pageMusic': {
+          import: './src/bandcamp/content/app.pageMusic.ts',
+          html: false,
+        },
         'bandcamp.content.album': {
           import: './src/bandcamp/content/app.album.ts',
           html: false,
